+++ src/js/pwa/register.js
@@ -2,18 +2,18 @@ import { showSnackbar } from '/js/ui/snackbar';
 
 export async function registerSW() {
   if (location.hostname === 'localhost' || location.hostname === '127.0.0.1' || location.hostname === '') {
-    // return false; // Disallow registering service worker on localhost
+    return false; // Disallow registering service worker on localhost
   }
 
   if ('serviceWorker' in navigator) {
     try {
-      navigator.serviceWorker.getRegistrations().then(function(registrations) {
+      /*navigator.serviceWorker.getRegistrations().then(function(registrations) {
         for (let i = 0; i < registrations.length; i++) {
           registrations[i].unregister().then(function(success) {
             showSnackbar("update");
           });
         }
-      });
+      });*/
 
       navigator.serviceWorker.register('/sw.js').then(reg => {
         reg.addEventListener('updatefound', () => {